Set property of gas tanks
